% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockpairwise-package.R
\docType{package}
\name{blockpairwise-package}
\alias{blockpairwise}
\alias{blockpairwise-package}
\title{blockpairwise}
\description{
Block pairwise composite likelihood tools.
}
\author{
\strong{Maintainer}: Bladimir Morales-Torrez \email{bvmorales@uc.cl}

Authors:
\itemize{
  \item Bladimir Morales-Torrez \email{bvmorales@uc.cl}
}

}
