% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.bp_fit}
\alias{coef.bp_fit}
\title{Extract model coefficients}
\usage{
\method{coef}{bp_fit}(object, ...)
}
\arguments{
\item{object}{Object of class `"bp_fit"`.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Numeric vector of estimated parameters.
}
\description{
Extract model coefficients
}
