% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_blockpairwise.R
\name{eval_blockpairwise}
\alias{eval_blockpairwise}
\title{Evaluate block-pairwise composite likelihood}
\usage{
eval_blockpairwise(par, bcl_data, type = c("marginal", "conditional"))
}
\arguments{
\item{par}{Numeric vector with parameters named `mu`,`sigma2`,`phi`,`tau2`, and `nu`.}

\item{bcl_data}{Object created by [prepare_bcl_data()].}

\item{type}{Character. Either `"marginal"` or `"conditional"`.}
}
\value{
Numeric value of the block-pairwise negative log-likelihood.
}
\description{
Evaluates the Gaussian block-pairwise composite negative log-likelihood
for a given parameter vector.
}
