% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.bp_fit}
\alias{logLik.bp_fit}
\title{Extract log-likelihood}
\usage{
\method{logLik}{bp_fit}(object, ...)
}
\arguments{
\item{object}{Object of class `"bp_fit"`.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Object of class `"logLik"`.
}
\description{
Extract log-likelihood
}
