% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_zi.R
\name{logLik.bp_zi_fit}
\alias{logLik.bp_zi_fit}
\title{Log-likelihood method for zero-inflated block-pairwise fit}
\usage{
\method{logLik}{bp_zi_fit}(object, ...)
}
\arguments{
\item{object}{object of class "bp_zi_fit".}

\item{...}{further arguments.}
}
\description{
Log-likelihood method for zero-inflated block-pairwise fit
}
