% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{make_pairs}
\alias{make_pairs}
\title{Creating pairs}
\usage{
make_pairs(coords, fweight = c("nn", "dist"), nnk = NULL, w = NULL)
}
\arguments{
\item{coords}{coordinates of the spatial points with columns x, y.}

\item{fweight}{pair construction method: "nn" or "dist".}

\item{nnk}{number of nearest neighbours, used if fweight = "nn".}

\item{w}{distance cut-off, used if fweight = "dist".}
}
\value{
Object of class list with pairs information.
}
\description{
Creating pairs
}
