% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.bp_fit}
\alias{predict.bp_fit}
\title{Predict from fitted block-pairwise model}
\usage{
\method{predict}{bp_fit}(object, newcoords, ...)
}
\arguments{
\item{object}{Object of class `"bp_fit"`.}

\item{newcoords}{Matrix of prediction coordinates.}

\item{...}{Further arguments.}
}
\description{
Predict from fitted block-pairwise model
}
