% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bp_fit}
\alias{print.bp_fit}
\title{Print method for bp_fit objects}
\usage{
\method{print}{bp_fit}(x, ...)
}
\arguments{
\item{x}{object of class bp_fit.}

\item{...}{ignored.}
}
\description{
Print method for bp_fit objects
}
