% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_zi.R
\name{print.bp_zi_fit}
\alias{print.bp_zi_fit}
\title{Print method for zero-inflated block-pairwise fit}
\usage{
\method{print}{bp_zi_fit}(x, ...)
}
\arguments{
\item{x}{object of class "bp_zi_fit".}

\item{...}{further arguments.}
}
\description{
Print method for zero-inflated block-pairwise fit
}
