% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.bp_fit}
\alias{print.summary.bp_fit}
\title{Print method for summary.bp_fit objects}
\usage{
\method{print}{summary.bp_fit}(x, ...)
}
\arguments{
\item{x}{object of class summary.bp_fit.}

\item{...}{ignored.}
}
\description{
Print method for summary.bp_fit objects
}
