% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.bp_precomp}
\alias{print.summary.bp_precomp}
\title{Print method for summary.bp_precomp objects}
\usage{
\method{print}{summary.bp_precomp}(x, ...)
}
\arguments{
\item{x}{object of class summary.bp_precomp.}

\item{...}{ignored.}
}
\description{
Print method for summary.bp_precomp objects
}
