% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.bp_fit}
\alias{simulate.bp_fit}
\title{Simulate from fitted block-pairwise model}
\usage{
\method{simulate}{bp_fit}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Object of class `"bp_fit"`.}

\item{nsim}{Number of simulations.}

\item{seed}{Optional random seed.}

\item{...}{Further arguments.}
}
\description{
Simulate from fitted block-pairwise model
}
