% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bp_fit}
\alias{summary.bp_fit}
\title{Summary method for bp_fit objects}
\usage{
\method{summary}{bp_fit}(object, ...)
}
\arguments{
\item{object}{object of class bp_fit.}

\item{...}{ignored.}
}
\description{
Summary method for bp_fit objects
}
