% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bp_precomp}
\alias{summary.bp_precomp}
\title{Summary method for bp_precomp objects}
\usage{
\method{summary}{bp_precomp}(object, ...)
}
\arguments{
\item{object}{object of class bp_precomp.}

\item{...}{ignored.}
}
\description{
Summary method for bp_precomp objects
}
