% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_zi.R
\name{coef.bp_zi_fit}
\alias{coef.bp_zi_fit}
\title{Extract coefficients from zero-inflated block-pairwise fit}
\usage{
\method{coef}{bp_zi_fit}(object, component = c("all", "mu", "zi", "dependence"), ...)
}
\arguments{
\item{object}{object of class "bp_zi_fit".}

\item{component}{one of "all", "mu", "zi", or "dependence".}

\item{...}{further arguments.}
}
\description{
Extract coefficients from zero-inflated block-pairwise fit
}
