% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{make_block}
\alias{make_block}
\title{Creating block labels}
\usage{
make_block(
  coords,
  cblocks = c("regular", "kmeans", "row", "col"),
  nblocks = 1L
)
}
\arguments{
\item{coords}{location coordinates.}

\item{cblocks}{block construction method.}

\item{nblocks}{number of spatial blocks.}
}
\value{
Object of class vector with block labels.
}
\description{
Creating block labels
}
