% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bcl_data.R
\name{prepare_bcl_data}
\alias{prepare_bcl_data}
\title{Prepare block-pairwise data}
\usage{
prepare_bcl_data(
  y,
  coords,
  X = NULL,
  cblocks = c("regular", "kmeans", "row", "col"),
  nblocks = 1L,
  fweight = c("nn", "dist"),
  nnk = NULL,
  w = NULL
)
}
\arguments{
\item{y}{numeric response vector.}

\item{coords}{numeric matrix/data.frame of spatial coordinates.}

\item{X}{optional design matrix. If NULL, an intercept-only model is used.}

\item{cblocks}{block construction method.}

\item{nblocks}{number of spatial blocks.}

\item{fweight}{pair construction method: "nn" or "dist".}

\item{nnk}{number of nearest neighbours, used if fweight = "nn".}

\item{w}{distance cut-off, used if fweight = "dist".}
}
\value{
An object of class `"bp_precomp"` containing block indices,
  within-block pairs, pairwise distances, block sizes and response values.
}
\description{
Prepare block-pairwise data
}
